requirements.txt
datasets==3.6.0
matplotlib==3.10.6
numpy==2.3.2
peft==0.17.0
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.7.1
torch==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.66.5
transformers==4.55.2
wandb==0.21.3
